added __SYSTEM__ subsection to default config
when the users config file was loaded, the __SYSTEM__ subsection was being
added at the time of validation.  It was not part of the template that was
written to disk.  This caused problems when writing the updated template
from within the UI.  ConfigObj was creating a __SYSTEM__ subsection but was
messing up its placement in the file.  The solution was to add an empty
__SYSTEM__ section to the default config template.