added ability to include external libraries
The SABnzbd zip archive that contains Cherrypy and Cheetah (Windows and OSX)
contains additional libraries that interfere with the successful loading of
the application.  The archive cannot be prepended to the include path (as is
done when setting the PYTHONPATH environment variable) and must be appended.
Therefore, the ui boot code was modified to accept a new command line flag
which allows users to register external libraries for inclusion.