bug fix: filesystem checks for relative paths
The new UI layout templates are stored local to the Media Rover application
directory and referred to in the config file by a relative filesystem path.
This caused problems with the config validation code (which expected an
absolute path).  Updated the check to first look for an absolute path and
on failure append the Media Rover application directory before trying again.