bug fix: fixed broken download cleanup
after Media Rover finished processingn a download, it tries to cleanup after
itself and delete the download directory.  Anything that can't be deleted is
moved to the .trash directory (in the first tv_root directory).  This was
failing if the download directory passed to the sort script ended with a /.
Updated the code to strip /, \, and ' ' off the end of the download path