bug fix: broken daily download/sorting for tvnzb
1. the tvnzb source plugin was trying to parse episode information from the
item description instead of the item title. In the case of daily episodes, tvnz
included the episode number (rather than the broadcast date) in the description
field.  This caused Media Rover to treat it like a series episode instead of a
daily.  Updated code to always use the title when gathering episode information

2. tvnzb names certain nzb's with a shortened series title instead of full title
according to tvrage.com.  This causes problems when a series folder was named
correctly (using full name) or when using multiple sources. Updated Media Rover
series templates to introduce the concept of the series alias.

closes #11