bug fix: parsing jobs in queue
the sabnzbd queue plugin already wrapped all attempts to create a newzbin
objects a try/catch.  It wasn't however doing the same for all other object
creation attempts.  This meant that whenever the queue tried to parse and
interpret a job that wasn't an episode and was from a source other than
newzbin, it would die.  Updated code to wrap ALL calls in a try/catch