bug fix to code that cleans up after sorting
there was a small bug in the code that attempts to clean up after the sorting
logic finishes.  It was attempting to create a folder in the .trash directory
then move the remaining download data to the same place.  This generated an
error because the destination folder already existed.  Removed mkdir logic
and allowed the move function to take care of the destination creation.