updated series to ignore files less than 50MB
updated series object to ignore files that are less than 50MB when building
the list of episodes found on disk.  This should prevent MR from accidentally
removing files that appear to be episodes (because they are named according
to a supported pattern) when in fact they are not (such as subtitle files).