refactored schedule command to check episode age
updated the schedule command to consult the filesystem and determine if a
potential download should be scheduled when the 'only_schedule_newer' flag
is set. It takes into account quality of existing episodes as well as
their relative age (broadcast date or episode number) in comparison to the
newest available series episode.