updated how MR manages queue
refactored the code responsible for managing the download queue. First, MR
now checks the queue for a matching job regardless of item delay. Second, MR
will no longer automatically stop jobs that are found to be less desirable
then a new offering. Instead, it will compare the size of the new item to
the remaining amount left to be downloaded. If the remaining amount is less
than the new item, the download will proceed and the new item will be given
a schedule delay of 1. In all other situations the queued job will be
immediately replaced.