refactored how Filesystem factory creates episodes
The Filesystem factory object was attempting to build a daily object before
a single object for all files found on disk. This behaviour likely dates back
to when the default single episode class supported a much broader (and more
difficult to parse) range of naming patterns. This caused a problem for a
single episode file with a title that had exactly 8 digits in a row which
were being identified as a date by the daily episode class.