updated filesystem to remove naming pattern
filesystem single and multipart episodes supported filenames without a season
and episode separator. This didn't present an issue in the past when the
fileystem factory attempted to create daily episodes before single or multipart
ones. However, a recent bug fix to address a fringe case (whereby an episode
title had 8 consecutive digits in a row) meant that daily episodes must
be created last. This change means the extra filesystem naming pattern can
and will improperly identify daily episodes as single episodes. Therefore, the
decision was made to remove support for this naming format, especially seeing
as it was only supported in filesystem episodes and not queue jobs or source
items.