refactored exceptions to remove second param
a second parameter was previous added to several exceptions that were generated
during the sort process. This was done so that when preparing the notification
messages, a formatted version of the episode object could be included. This
proved problematic as the episode object is not available in the method that
initiates the sort. Refactored code to reverse addition of extra param